Refactor service classes
