+++ src/DevOpsAssistant/DevOpsAssistant/Services/Models/WorkItemInfo.cs
@@ -1,0 +1,10 @@
+namespace DevOpsAssistant.Services;
+
+public class WorkItemInfo
+{
+    public int Id { get; set; }
+    public string Title { get; set; } = string.Empty;
+    public string State { get; set; } = string.Empty;
+    public string WorkItemType { get; set; } = string.Empty;
+    public string Url { get; set; } = string.Empty;
+}
